whole time step in calculer_indicatrice_next(Ã)

+++ Multiphase/Front_tracking_IJK/src/FT/IJK_FT.h
@@ -83,7 +83,7 @@ public :
   {
     return current_time_ ;
   }
-  const int& get_timestep() const
+  const double& get_timestep() const
   {
     return timestep_ ;
   }
